meteodata/requirements.txt
## requirements for downloading files
requests
python-dotenv
pytest
fastapi
uvicorn
google-cloud-storage